refactor: Update Model component with dynamic classes
The Model component has been updated to include dynamic classes. This allows for more flexibility in styling the modal. The `classes` prop has been added to the component, which can be used to pass in additional CSS classes for customization. This change improves the reusability and customization options of the Model component.